Extended: ensure all fixers checking disposal state

+++ src/main/java/com/kalessil/phpStorm/phpInspectionsEA/inspectors/codeStyle/UnusedGotoLabelInspector.java
@@ -84,7 +84,7 @@ public class UnusedGotoLabelInspector extends BasePhpInspection {
         @Override
         public void applyFix(@NotNull Project project, @NotNull ProblemDescriptor descriptor) {
             final PsiElement target = descriptor.getPsiElement();
-            if (target instanceof PhpGotoLabel) {
+            if (target instanceof PhpGotoLabel && !project.isDisposed()) {
                 target.delete();
             }
         }
